feat: associate crops and fertilizers and get crop fertilizers

+++ src/main/java/com/betrybe/agrix/dto/FertilizerDto.java
@@ -3,6 +3,6 @@ package com.betrybe.agrix.dto;
 /**
  * FertilizerDto.
  */
-public record FertilizerDto(Long id, String name, String brand, String composition) {
+public record FertilizerDto(Integer id, String name, String brand, String composition) {
     
 }
