+++ src/main/java/com/betrybe/agrix/repository/CropRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Crop Repository.
  */
-public interface CropRepository extends JpaRepository<Crop, Long> {
+public interface CropRepository extends JpaRepository<Crop, Integer> {
     
 }
